# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  apierror.ErrorResponse:
    properties:
      error:
        description: Stable error code (e.g., "invalid_credentials")
        type: string
      message:
        description: Human-readable message
        type: string
    type: object
  dto.AddRepositoryRequest:
    properties:
      display_name:
        maxLength: 255
        minLength: 1
        type: string
      git_credential_id:
        type: string
      git_default_branch:
        type: string
      git_integration_credential_id:
        description: GitIntegrationCredentialID — выбранный OAuth-аккаунт провайдера
          для этого репо (мульти-аккаунт).
        type: string
      git_provider:
        type: string
      git_url:
        type: string
      is_primary:
        type: boolean
      role_description:
        type: string
      slug:
        maxLength: 64
        minLength: 1
        type: string
      sort_order:
        type: integer
    required:
    - display_name
    - git_url
    - slug
    type: object
...
  dto.GitCredentialSummary:
    properties:
      auth_type:
        type: string
      id:
        type: string
      label:
        type: string
      provider:
        type: string
    type: object
...
  dto.ProjectRepositoryListResponse:
    properties:
      repositories:
        items:
          $ref: '#/definitions/dto.ProjectRepositoryResponse'
        type: array
      total:
        type: integer
    type: object
  dto.ProjectRepositoryResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      git_credential:
        $ref: '#/definitions/dto.GitCredentialSummary'
      git_default_branch:
        type: string
      git_integration_credential_id:
        type: string
      git_provider:
        type: string
      git_url:
        type: string
      id:
        type: string
      is_primary:
        type: boolean
      last_indexed_commit:
        type: string
      project_id:
        type: string
      role_description:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      vector_collection:
        type: string
    type: object
...
  dto.UpdateRepositoryRequest:
    properties:
      display_name:
        type: string
      git_credential_id:
        type: string
      git_default_branch:
        type: string
      git_integration_credential_id:
        type: string
      git_provider:
        type: string
      git_url:
        type: string
      is_primary:
        type: boolean
      remove_git_integration_credential:
        type: boolean
      role_description:
        type: string
      sort_order:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Backend API с авторизацией на JWT токенах
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /projects/{id}/repositories:
    get:
      description: Возвращает git-репозитории проекта (мульти-репо).
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProjectRepositoryListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      - OAuth2Password: []
      summary: Список репозиториев проекта
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Добавляет git-репозиторий в проект (мульти-репо). Первый репозиторий
        становится primary.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Данные репозитория
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddRepositoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProjectRepositoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      - OAuth2Password: []
      summary: Добавление репозитория в проект
      tags:
      - projects
  /projects/{id}/repositories/{repoId}:
    delete:
      description: Удаляет git-репозиторий из проекта. Primary нельзя удалить, пока
        есть другие репозитории.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Repository ID
        in: path
        name: repoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      - OAuth2Password: []
      summary: Удаление репозитория проекта
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Частичное обновление репозитория проекта (мульти-репо).
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Repository ID
        in: path
        name: repoId
        required: true
        type: string
      - description: Поля для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRepositoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProjectRepositoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      - OAuth2Password: []
      summary: Обновление репозитория проекта
      tags:
      - projects
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Long-lived API key for programmatic access. Format: wibe_<key>'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  OAuth2Password:
    flow: password
    tokenUrl: /api/v1/auth/login
    type: oauth2
swagger: "2.0"